chore: Update environment variables in build.yml
This commit updates the environment variables in the `.github/workflows/build.yml` file. The `STACK_NAME` is set to `express-bdoilerplate`, `DOMAIN` is set to `localhost`, `API_KEY` is set to `'secret'`, and `SESSION_SECRET` is set to `'secret'`. These changes ensure that the correct values are used for the development environment.