chore: Update packageManager to pnpm@9.7.0 and dependencies
This commit updates the packageManager field in the package.json file to use pnpm@9.7.0 instead of pnpm@9.6.0. It also updates the dependencies in the project, including adding the "rotating-file-stream" plugin with version 3.2.3. These changes ensure that the project is using the latest stable versions of the package manager and dependencies.