chore: Update .dockerignore and .prettierignore files, and import types in user.repository.ts and service-response.model.ts
This commit updates the .dockerignore file to include "*.log" as an ignored file. It also removes the .prettierignore file, as it is no longer needed. In the user.repository.ts file, the imports for the User type are updated to use the "type" keyword. Similarly, in the service-response.model.ts file, the enum values for ResponseStatus are updated to use explicit numeric values. These changes improve the project's file management and type handling.